pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "inner-heart-ai-chat"
version = "0.1.0"
description = "속마음 말하기: AI-powered empathetic chat with multi-agent architecture"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name"}
]
keywords = ["ai", "chat", "psychology", "multi-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications",
]

dependencies = [
    "fastapi==0.120.0",
    "uvicorn[standard]==0.31.1",
    "pydantic==2.10.3",
    "pydantic-settings==2.3.1",
    "python-dotenv==1.0.1",
    "requests==2.31.0",
    "httpx==0.27.0",
    "aiofiles==23.2.1",
    "typing-extensions==4.12.2",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
    "pytest-cov==4.1.0",
    "black==24.1.1",
    "ruff==0.2.2",
    "mypy==1.10.0",
    "isort==5.13.2",
    "pre-commit==3.7.1",
]
rag = [
    "scikit-learn==1.5.0",
    "sentence-transformers==3.0.1",
    "numpy==1.26.4",
    "faiss-cpu==1.8.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/inner-heart-ai-chat"
Repository = "https://github.com/yourusername/inner-heart-ai-chat.git"
Issues = "https://github.com/yourusername/inner-heart-ai-chat/issues"

[tool.uv]
python-version = "3.11"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # 디렉토리
  \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = [".venv", "dist", "build"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = false
check_untyped_defs = true

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",
    "W",
    "F",
    "I",
    "N",
    "UP",
    "YTT",
    "B",
    "A",
    "C4",
    "T10",
    "T20",
    "SIM",
    "ERA",
    "D100",
    "D101",
    "D102",
]
ignore = [
    "E501",  # line-too-long (black이 처리)
    "D104",  # missing-docstring-in-public-package
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "--cov=src --cov-report=html --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
]
